requirements.txt
# Core dependencies for WAT framework tools

# Environment management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Data processing
pandas==2.1.4

# Google API (if using Google Sheets, Drive, etc.)
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Web scraping (if needed)
beautifulsoup4==4.12.2
lxml==5.0.0

# Web application (LinkedIn Posts Scraper)
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Deliverability Dashboard
flask==3.0.3
gunicorn==22.0.0

# Add other dependencies as needed for your specific tools
